pcn-k8s: enforce in a different goroutine
If the number of rules to push exceeds a certain number,
a new goroutine is created for them. There is no real
benefit in launching new goroutines when the number
of rules are very few, as the cost of launching a new
goroutine (even if they are light threads) could be
too high.

Signed-off-by: Elis Lulja <elis.lulja@gmail.com>